· Order of Includes ( http://google-styleguide.googlecode.com/svn/trunk/cppguide.html#Names_and_Order_of_Includes )
1. dir2/foo2.h.
2. C system files.
3. C++ system files.
4. Other libraries' .h files.
5. Your project's .h files.

+++ SceneR.Framework/Headers/System/Core.hpp
@@ -33,7 +33,7 @@ namespace System
     using UInt64  = uint64_t;
     using Single  = float;
     using Double  = double;
-    using Size    = size_t;
+    using Size    = uint32_t;
     using String  = std::u16string;
 }
